Handle set-debugger-loc for Linux/Windows
* just open socket server listen message, so the app cannot automatically open
